fix faker number type

+++ SmartClinic-Server/Tests/unit/createVital.test.js
@@ -1,5 +1,5 @@
 import request from 'supertest'
-import app from '../../app.js' // your Express app
+import app from '../../app.js' 
 import { faker } from '@faker-js/faker'
 
 describe('POST /doctor/createvital', () => {
@@ -7,10 +7,10 @@ describe('POST /doctor/createvital', () => {
       const payload = {
       patientId: 2, // Use existing test data
       doctorId: 1,
-      healthPercentage: faker.number.float({ min: 50, max: 100, precision: 0.1 }),
+      healthPercentage: faker.number.int({ min: 50, max: 100, precision: 0.1 }),
       heartRate: faker.number.int({ min: 60, max: 120 }),
       bloodPressure: `${faker.number.int({ min: 100, max: 140 })}/${faker.number.int({ min: 60, max: 90 })}`,
-      temperature: faker.number.float({ min: 36.0, max: 38.0, precision: 0.1 }),
+      temperature: faker.number.int({ min: 36.0, max: 38.0, precision: 0.1 }),
       bloodGlucose: faker.number.int({ min: 70, max: 150 }),
     };
     const res = await request(app)
